Cleaned up security stuff
