+++ .gitignore
@@ -24,3 +24,4 @@
 /public/assets/
 /assets/vendor/
 ###< symfony/asset-mapper ###
+*.local
