+++ api/README.md
@@ -13,7 +13,6 @@ GET /api/v1/process/{id}/relations/run
 (GET /api/v1/process/{id}/run/{id})
 POST /api/v1/run/{id}/retry (/api/v1/process/{id}/run/{id}/retry)
 (GET /api/v1/process/{id}/run/search)
-
 ```
 
 <https://fastapi.tiangolo.com/virtual-environments/#create-a-virtual-environment>
@@ -53,4 +52,23 @@ curl --silent --verbose --location 'http://127.0.0.1:8000/api/v1/process/1/run/3
     "started_at": "2025-09-25"
 }
 '
+```
+
+## Security
+
+Defined API keys in `.env.local`, e.g.
+
+``` dotenv
+# .env.local
+# Get a token from https://generate-random.org/api-token-generator or some such …
+API_KEYS_READ='["a-not-so-secret-key", "759568492f338454603821a04810eabf"]'
+API_KEYS_WRITE='["3825e7be2d1ca130063171d8362ad4996e3a0df1e9f6dd2a4dc6bebf38bfc205"]'
+```
+
+Restart the API to load the updated config.
+
+Use a key:
+
+``` shell
+curl http://127.0.0.1:8000/api/v1/process/ --header 'x-api-key: a-not-so-secret-key'
 ```
